requirements.txt
mcp
streamlit
openai
python-dotenv
httpx
httpx-sse
